Move rsc stuff to entry-client
